support for all passbook images
Added support for all images (icon, logo, background, thumbnail and
strip).

In order to give support to all documented images, and instead of
creating objects for each file, as the images are optional, one created
a simple object that holds the Images Path and all you need is to drop
all styled images (following the documented file names and sizes - link
is provided) into that folder.

ImagesList holds the possibility to override each file on that folder.
Example is shown on the EventTicket Action.