replace System.Drawing dep with simple color parsing
The System.Drawing.Common package has dependencies
on win32 specific packages. I didn't test the `ColorTranslator`
on Linux, but thought it would be best to replace the it and
remove a potential cross-platform compatibility problem.

Initial tests with my application on Linux look good. Packages
generate works signing appears to be valid.